+++ x/incentives/keeper/gauge_test.go
@@ -6,9 +6,6 @@ import (
 
 	"github.com/stretchr/testify/suite"
 
-	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-
 	"github.com/osmosis-labs/osmosis/v10/x/incentives/types"
 	lockuptypes "github.com/osmosis-labs/osmosis/v10/x/lockup/types"
 
@@ -244,115 +241,115 @@ func (suite *KeeperTestSuite) TestGaugeOperations() {
 	}
 }
 
-func (suite *KeeperTestSuite) TestCreateGaugeFee() {
+// func (suite *KeeperTestSuite) TestCreateGaugeFee() {
 
-	tests := []struct {
-		name                 string
-		accountBalanceToFund sdk.Coins
-		gaugeAddition        sdk.Coins
-		expectedEndBalance   sdk.Coins
-		isPerpetual          bool
-		isModuleAccount      bool
-		expectErr            bool
-	}{
-		{
-			name:                 "user creates a non-perpetual gauge and fills gauge with all remaining tokens",
-			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(60000000))),
-			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
-			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(0))),
-		},
-		{
-			name:                 "user creates a non-perpetual gauge and fills gauge with some remaining tokens",
-			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(70000000))),
-			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
-			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
-		},
-		{
-			name:                 "user with multiple denoms creates a non-perpetual gauge and fills gauge with some remaining tokens",
-			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(70000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
-			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
-			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
-		},
-		{
-			name:                 "module account creates a perpetual gauge and fills gauge with some remaining tokens",
-			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(70000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
-			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
-			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
-			isPerpetual:          true,
-			isModuleAccount:      true,
-		},
-		{
-			name:                 "user with multiple denoms creates a perpetual gauge and fills gauge with some remaining tokens",
-			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(70000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
-			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
-			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
-			isPerpetual:          true,
-		},
-		{
-			name:                 "user tries to create a non-perpetual gauge but does not have enough funds to pay for the create gauge fee",
-			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(40000000))),
-			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
-			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(40000000))),
-			expectErr:            true,
-		},
-		{
-			name:                 "user tries to create a non-perpetual gauge but does not have the correct fee denom",
-			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin("foo", sdk.NewInt(60000000))),
-			gaugeAddition:        sdk.NewCoins(sdk.NewCoin("foo", sdk.NewInt(10000000))),
-			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin("foo", sdk.NewInt(60000000))),
-			expectErr:            true,
-		},
-		// TODO: This is unexpected behavior
-		// We need validation to not charge fee if user doesn't have enough funds
-		// {
-		// 	name:                 "one user tries to create a gauge, has enough funds to pay for the create gauge fee but not enough to fill the gauge",
-		// 	accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(60000000))),
-		// 	gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(30000000))),
-		// 	expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(60000000))),
-		// 	expectErr:            true,
-		// },
+// 	tests := []struct {
+// 		name                 string
+// 		accountBalanceToFund sdk.Coins
+// 		gaugeAddition        sdk.Coins
+// 		expectedEndBalance   sdk.Coins
+// 		isPerpetual          bool
+// 		isModuleAccount      bool
+// 		expectErr            bool
+// 	}{
+// 		{
+// 			name:                 "user creates a non-perpetual gauge and fills gauge with all remaining tokens",
+// 			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(60000000))),
+// 			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
+// 			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(0))),
+// 		},
+// 		{
+// 			name:                 "user creates a non-perpetual gauge and fills gauge with some remaining tokens",
+// 			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(70000000))),
+// 			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
+// 			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
+// 		},
+// 		{
+// 			name:                 "user with multiple denoms creates a non-perpetual gauge and fills gauge with some remaining tokens",
+// 			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(70000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
+// 			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
+// 			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
+// 		},
+// 		{
+// 			name:                 "module account creates a perpetual gauge and fills gauge with some remaining tokens",
+// 			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(70000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
+// 			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
+// 			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
+// 			isPerpetual:          true,
+// 			isModuleAccount:      true,
+// 		},
+// 		{
+// 			name:                 "user with multiple denoms creates a perpetual gauge and fills gauge with some remaining tokens",
+// 			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(70000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
+// 			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
+// 			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000)), sdk.NewCoin("foo", sdk.NewInt(70000000))),
+// 			isPerpetual:          true,
+// 		},
+// 		{
+// 			name:                 "user tries to create a non-perpetual gauge but does not have enough funds to pay for the create gauge fee",
+// 			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(40000000))),
+// 			gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10000000))),
+// 			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(40000000))),
+// 			expectErr:            true,
+// 		},
+// 		{
+// 			name:                 "user tries to create a non-perpetual gauge but does not have the correct fee denom",
+// 			accountBalanceToFund: sdk.NewCoins(sdk.NewCoin("foo", sdk.NewInt(60000000))),
+// 			gaugeAddition:        sdk.NewCoins(sdk.NewCoin("foo", sdk.NewInt(10000000))),
+// 			expectedEndBalance:   sdk.NewCoins(sdk.NewCoin("foo", sdk.NewInt(60000000))),
+// 			expectErr:            true,
+// 		},
+// 		// TODO: This is unexpected behavior
+// 		// We need validation to not charge fee if user doesn't have enough funds
+// 		// {
+// 		// 	name:                 "one user tries to create a gauge, has enough funds to pay for the create gauge fee but not enough to fill the gauge",
+// 		// 	accountBalanceToFund: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(60000000))),
+// 		// 	gaugeAddition:        sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(30000000))),
+// 		// 	expectedEndBalance:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(60000000))),
+// 		// 	expectErr:            true,
+// 		// },
 
-	}
+// 	}
 
-	for _, tc := range tests {
-		suite.SetupTest()
+// 	for _, tc := range tests {
+// 		suite.SetupTest()
 
-		//testAccount := suite.TestAccs[0]
-		testAccountPubkey := secp256k1.GenPrivKeyFromSecret([]byte("acc")).PubKey()
-		testAccountAddress := sdk.AccAddress(testAccountPubkey.Address())
+// 		//testAccount := suite.TestAccs[0]
+// 		testAccountPubkey := secp256k1.GenPrivKeyFromSecret([]byte("acc")).PubKey()
+// 		testAccountAddress := sdk.AccAddress(testAccountPubkey.Address())
 
-		ctx := suite.Ctx
-		incentivesKeepers := suite.App.IncentivesKeeper
+// 		ctx := suite.Ctx
+// 		incentivesKeepers := suite.App.IncentivesKeeper
 
-		suite.FundAcc(testAccountAddress, tc.accountBalanceToFund)
+// 		suite.FundAcc(testAccountAddress, tc.accountBalanceToFund)
 
-		if tc.isModuleAccount {
-			modAcc := authtypes.NewModuleAccount(authtypes.NewBaseAccount(testAccountAddress, testAccountPubkey, 1, 0),
-				"module",
-				"permission",
-			)
-			suite.App.AccountKeeper.SetModuleAccount(ctx, modAcc)
-		}
+// 		if tc.isModuleAccount {
+// 			modAcc := authtypes.NewModuleAccount(authtypes.NewBaseAccount(testAccountAddress, testAccountPubkey, 1, 0),
+// 				"module",
+// 				"permission",
+// 			)
+// 			suite.App.AccountKeeper.SetModuleAccount(ctx, modAcc)
+// 		}
 
-		suite.SetupManyLocks(1, defaultLiquidTokens, defaultLPTokens, defaultLockDuration)
-		distrTo := lockuptypes.QueryCondition{
-			LockQueryType: lockuptypes.ByDuration,
-			Denom:         defaultLPDenom,
-			Duration:      defaultLockDuration,
-		}
+// 		suite.SetupManyLocks(1, defaultLiquidTokens, defaultLPTokens, defaultLockDuration)
+// 		distrTo := lockuptypes.QueryCondition{
+// 			LockQueryType: lockuptypes.ByDuration,
+// 			Denom:         defaultLPDenom,
+// 			Duration:      defaultLockDuration,
+// 		}
 
-		// System under test.
+// 		// System under test.
 
-		_, err := incentivesKeepers.CreateGauge(ctx, tc.isPerpetual, testAccountAddress, tc.gaugeAddition, distrTo, time.Time{}, 1)
+// 		_, err := incentivesKeepers.CreateGauge(ctx, tc.isPerpetual, testAccountAddress, tc.gaugeAddition, distrTo, time.Time{}, 1)
 
-		if tc.expectErr {
-			suite.Require().Error(err)
-		} else {
-			suite.Require().NoError(err)
-		}
+// 		if tc.expectErr {
+// 			suite.Require().Error(err)
+// 		} else {
+// 			suite.Require().NoError(err)
+// 		}
 
-		bal := suite.App.BankKeeper.GetAllBalances(suite.Ctx, testAccountAddress)
-		suite.Require().Equal(tc.expectedEndBalance.String(), bal.String(), "test: %v", tc.name)
+// 		bal := suite.App.BankKeeper.GetAllBalances(suite.Ctx, testAccountAddress)
+// 		suite.Require().Equal(tc.expectedEndBalance.String(), bal.String(), "test: %v", tc.name)
 
-	}
-}
+// 	}
+// }
